Data Engineer/dim_country/requirements.txt
pandas
sqlalchemy
python-dotenv
mysql.connector
os